+++ pmma/core/cpp_src/Animation/RadialAnimation.cpp
@@ -42,7 +42,7 @@ void CPP_RadialAnimation::Start() {
     StartTime = chrono::high_resolution_clock::now();
     RunTime = chrono::seconds(0);
 
-    unsigned int start_position[2];
+    float start_position[2];
     StartCoordinatePtr->Get(start_position);
     TargetCoordinatePtr->Set(start_position);
 }
